refactor: dashboard components folder names
