+++ packages/client/components/dashboard/header/DashboardHeader.tsx
@@ -1,9 +1,9 @@
 import { Sheet, SheetContent, SheetTrigger } from '@client/components/ui/sheet';
 import { Menu } from 'lucide-react';
-import DashboardSidebar from '../dashboard-sidebar/DashboardSidebar';
 import { cn } from '@client/client-utils/tailwind/cn';
 import TeacherModeButton from '@client/components/teacher-mode-button/TeacherModeButton';
 import ClerkUserButton from '@client/components/clerk-user-button/ClerkUserButton';
+import DashboardSidebar from '../sidebar/DashboardSidebar';
 
 const DashboardHeader = () => {
   return (
